# web-ui/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AIMS Web Application API
  description: |
    Complete API specification for the AIMS (Internet Media Store) Web Application.
    
    This API provides comprehensive e-commerce functionality including:
    - Product catalog management
    - Shopping cart operations
    - Order processing and tracking
    - Payment integration with VNPay
    - User authentication and management
    - Administrative operations
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer your_jwt_token_here
    ```
    
    ## Rate Limiting
    API requests are rate limited per user:
    - General endpoints: 1000 requests/hour
    - Authentication: 10 requests/minute
    - File uploads: 100 requests/hour
    
  version: 1.0.0
  contact:
    name: AIMS Development Team
    email: dev@aims.com
  license:
    name: Proprietary
    url: https://aims.com/license
servers:
  - url: https://api.aims.com
    description: Production server
  - url: https://staging-api.aims.com
    description: Staging server
  - url: http://localhost:8080/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get products
      description: Retrieve paginated list of products with optional filtering
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: productType
          in: query
          schema:
            type: string
            enum: [Book, CD, DVD, LP]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
  /products/{productId}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve detailed information about a specific product
      security: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Cart Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          example: prod_123
        title:
          type: string
          example: "The Great Gatsby"
        price:
          type: number
          format: float
          example: 15.99
        category:
          type: string
          example: "Classic Literature"
        productType:
          type: string
          enum: [Book, CD, DVD, LP]
          example: Book
        author:
          type: string
          example: "F. Scott Fitzgerald"
        description:
          type: string
          example: "A classic American novel..."
        imageUrl:
          type: string
          example: "https://images.aims.com/products/great-gatsby.jpg"
        stockQuantity:
          type: integer
          example: 25
        isAvailable:
          type: boolean
          example: true
        specifications:
          type: object
          additionalProperties: true
    ProductListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Products retrieved successfully
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ProductResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Product retrieved successfully
        data:
          $ref: '#/components/schemas/Product'

    # Cart Schemas
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              code:
                type: string
                example: INVALID_EMAIL
              message:
                type: string
                example: Please provide a valid email address
...